// src/main/frontend/src/services/openapi/models/EntryFullDTO.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { EntryDTO } from './EntryDTO';
import type { ImageDTO } from './ImageDTO';
export type EntryFullDTO = {
    object?: EntryDTO;
    domObjects?: EntryDTO;
    subObjects?: Array<EntryDTO>;
    images?: Array<ImageDTO>;
};